match/requirements.txt
redis
psycopg2-binary
sqlalchemy
aiohttp
asyncio
python-dotenv
httpx
loguru
pydantic 